Get music name from sroucecode
Radio Javan has emerged in a new design so the scraping methods need to
be changed. In the new source code, there is an element with the id of
`__NEXT_DATA__` containing the media data along side related tracks,
user-agent, etc in a js object. So to grab the info we will use this
object instead of scraping through the website (which is impossible with
the way the code currently works since rj is leaning towards frontend
rendering)

 - Add mock source-codes as html files since tests will no more rely on
   small pieces of source-code
 - Get music name from the raw data that RadioJavan presents in the page