Add `user-agent` to request headers
This change comes from a field represented in `__NEXT_DATA__` element.
The field is called `isBot` and it's a boolean. If we send a request
with normal axios headers it'll become true but if we simply use a
browser-like header it'll be false. Also the user-agent is represented
in the element as well. So I assume it's it's better to override the
default headers with a browser-like one