Update url type utilities
Radiojavan has a new url format which not only includes a new origin
(`play.radiojavan.com` instead of `radiojavan.com`) but new routings as
well. So in order for link types to be accurate, changes must be made
upon regexes. Also url formatter needs to acclimate to the new format

 - Update url formatter regex ('radioJavanLinkRegex')
 - Update url type regexes ('typeRegexes')
 - Update 'formatURL' function to use the new origin
 - Delete `PlaylistTrack` and `AlbumTrack` from extended types since
   they aren't a thing in the new url shape